refactor(ml-1268): extract loadQuestion helper for question 404 path
questionController and questionPostController duplicated the
look-up-or-404 block verbatim apart from a GET/POST verb in the log
message. Move it into question/utils.js as loadQuestion(request),
returning { questionRoute, question } on success and
logging+throwing Boom.notFound on miss. Drop the verb prefix from the
fix/summary strings — the action label unknown-question-route is
verb-agnostic, ECS auto-captures http.request.method, and the route
is the triage key. Add a missing POST-side parity test asserting the
warn fires on 404 (the GET-side test already had this).
