+++ src/server/journey/self-service/question/utils.js
@@ -1,3 +1,25 @@
+import Boom from '@hapi/boom'
+import { getQuestion } from '#src/server/journey/self-service/services/journey-data.js'
+import { reportRuntimeIssue } from '#src/server/journey/self-service/services/data-quality.js'
+
 export function questionRouteFromRequest(request) {
   return '/' + request.params.questionPath
+}
+
+export function loadQuestion(request) {
+  const questionRoute = questionRouteFromRequest(request)
+  const question = getQuestion(questionRoute)
+
+  if (!question) {
+    reportRuntimeIssue(
+      request,
+      'unknown-question-route',
+      questionRoute,
+      `Add ${questionRoute} as a question route or fix the referring answer in self-service.json`,
+      `unknown question route ${questionRoute}`
+    )
+    throw Boom.notFound('Question not found')
+  }
+
+  return { questionRoute, question }
 }
